pyproject.toml
[project]
name = "archaeologist-prototype"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">3.10"
dependencies = [
    "demucs>=4.0.1",
    "fastapi>=0.116.1",
    "faster-whisper>=1.2.0",
    "librosa>=0.11.0",
    "numpy>=2.2.6",
    "openai-whisper>=20250625",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "soundfile>=0.13.1",
    "torch>=2.8.0",
    "tqdm>=4.67.1",
    "uvicorn>=0.35.0",
    "whisperx>=3.4.2",
]
